+++ resources/views/mark/edit.blade.php
@@ -416,17 +416,17 @@
                 $parent.find('#grade').text("F");
             });
 
-            var pass_written = parseInt($('#pass_written').text()) || 0;
-            var pass_mcq = parseInt($('#pass_mcq').text()) || 0;
-            var pass_practical = parseInt($('#pass_practical').text()) || 0;
+            var pass_written = parseFloat($('#pass_written').text()) || 0;
+            var pass_mcq = parseFloat($('#pass_mcq').text()) || 0;
+            var pass_practical = parseFloat($('#pass_practical').text()) || 0;
 
 
             $('.ts-student-row input').on('keyup', function () {
                 $parent = $(this).parent().parent();
 
-                $written = parseInt($parent.find('#written').val()) || 0;
-                $mcq = parseInt($parent.find('#mcq').val()) || 0;
-                $practical = parseInt($parent.find('#practical').val()) || 0;
+                $written = parseFloat($parent.find('#written').val()) || 0;
+                $mcq = parseFloat($parent.find('#mcq').val()) || 0;
+                $practical = parseFloat($parent.find('#practical').val()) || 0;
 
                 var is_fail = false;
 
@@ -455,9 +455,9 @@
             $('.ts-student-row input#absent').each(function () {
                 $parent = $(this).parent().parent();
 
-                $written = parseInt($parent.find('#written').val()) || 0;
-                $mcq = parseInt($parent.find('#mcq').val()) || 0;
-                $practical = parseInt($parent.find('#practical').val()) || 0;
+                $written = parseFloat($parent.find('#written').val()) || 0;
+                $mcq = parseFloat($parent.find('#mcq').val()) || 0;
+                $practical = parseFloat($parent.find('#practical').val()) || 0;
 
                 var is_fail = false;
 
@@ -504,7 +504,7 @@
         ];
 
         function getPoint(val) {
-            val = parseInt(val);
+            val = parseFloat(val);
 
             var length = grade_value.length;
 
@@ -520,7 +520,7 @@
         }
 
         function getGrade(val) {
-            val = parseInt(val);
+            val = parseFloat(val);
 
             var length = grade_value.length;
